fix delete admin

+++ src/controllers/adminController.js
@@ -101,17 +101,22 @@ module.exports.createAdmin = asyncHandler(async (req, res) => {
  *---------------------------------*/
 module.exports.deleteAdmin = asyncHandler(async (req, res) => {
   const { id } = req.params
-
-  if (!id) {
-    return res.status(400).json({ message: 'ID parameter is required' })
+  const currentAdminId = req.admin.id // Assuming req.user.id contains the ID of the authenticated admin
+ if (!id) {
+   return res.status(400).json({ message: 'ID parameter is required' })
+ }
+  
+  if (id === currentAdminId) {
+    return res
+      .status(403)
+      .json({ message: 'You cannot delete your own account' })
   }
-
+   
   let admin = await Admin.findById(id)
 
   if (!admin) {
-    return res.status(400).json({ message: "this admin doesn't exist" })
+    return res.status(400).json({ message: "This admin doesn't exist" })
   }
-
   await Admin.findByIdAndDelete(admin.id)
 
   return res.status(200).json({ message: 'Admin deleted successfully' })
